refactor: Shader parameters source of truth is host
This change removes the dependency for loading and retrieving shader
parameter values from the helper. This simplifies the dataflow,
as it only flows down to the helper rather than bi-directional.

This is foundational work to improve the shader experiece. Future
iterations will include support for sharing presets, more flexible
UI customisation and more.